multer middelware and cloudinary setup

+++ Backend/server/config/cloudinary.js
@@ -2,7 +2,7 @@ import {v2 as cloudinary } from 'cloudinary';
 import fs from 'fs';
 
 
-cloudinary.comfig({
+cloudinary.config({
     cloud_name : "ProtoVolt",
     api_key: process.env.cloudinary_key,
     api_secret: process.env.cloudinary_secret
@@ -14,11 +14,13 @@ const uploadCloudinary = async (localFilePath)=>{
         if(!localFilePath) return { success:false ,message:"No file Tracked"};
 
         const res = await cloudinary.uploader.upload(localFilePath,{
+            folder,
             resource_type:'auto'
         });
+        if(fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return res;
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        if(fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return { success:false ,message:"Something happen Reupload Your File"};
     }
 }
